backend/requirements.txt
pika
psycopg2-binary
django
python-dotenv
pytest
pytest-django
djangorestframework
factory-boy
